Refactor step handler to a standalone StepRequestHandler
