+++ com.microsoft.java.debug.core/src/main/java/com/microsoft/java/debug/core/adapter/handler/ThreadsRequestHandler.java
@@ -13,14 +13,10 @@ package com.microsoft.java.debug.core.adapter.handler;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 
-import com.microsoft.java.debug.core.DebugEvent;
 import com.microsoft.java.debug.core.DebugUtility;
-import com.microsoft.java.debug.core.IDebugSession;
 import com.microsoft.java.debug.core.adapter.AdapterUtils;
 import com.microsoft.java.debug.core.adapter.ErrorCode;
 import com.microsoft.java.debug.core.adapter.IDebugAdapterContext;
@@ -31,31 +27,19 @@ import com.microsoft.java.debug.core.protocol.Requests;
 import com.microsoft.java.debug.core.protocol.Requests.Arguments;
 import com.microsoft.java.debug.core.protocol.Requests.Command;
 import com.microsoft.java.debug.core.protocol.Requests.ContinueArguments;
-import com.microsoft.java.debug.core.protocol.Requests.NextArguments;
 import com.microsoft.java.debug.core.protocol.Requests.PauseArguments;
-import com.microsoft.java.debug.core.protocol.Requests.StepInArguments;
-import com.microsoft.java.debug.core.protocol.Requests.StepOutArguments;
 import com.microsoft.java.debug.core.protocol.Requests.ThreadsArguments;
 import com.microsoft.java.debug.core.protocol.Responses;
 import com.microsoft.java.debug.core.protocol.Types;
-import com.sun.jdi.Location;
-import com.sun.jdi.Method;
 import com.sun.jdi.ObjectCollectedException;
-import com.sun.jdi.StackFrame;
 import com.sun.jdi.ThreadReference;
 import com.sun.jdi.VMDisconnectedException;
-import com.sun.jdi.event.StepEvent;
-import com.sun.jdi.request.StepRequest;
-
-import io.reactivex.disposables.Disposable;
 
 public class ThreadsRequestHandler implements IDebugRequestHandler {
-    private Disposable eventHandler = null;
-    private Map<Long, ThreadState> threadStates = new HashMap<>();
 
     @Override
     public List<Command> getTargetCommands() {
-        return Arrays.asList(Command.THREADS, Command.STEPIN, Command.STEPOUT, Command.NEXT, Command.PAUSE, Command.CONTINUE);
+        return Arrays.asList(Command.THREADS, Command.PAUSE, Command.CONTINUE);
     }
 
     @Override
@@ -64,21 +48,9 @@ public class ThreadsRequestHandler implements IDebugRequestHandler {
             return AdapterUtils.createAsyncErrorResponse(response, ErrorCode.EMPTY_DEBUG_SESSION, "Debug Session doesn't exist.");
         }
 
-        if (eventHandler == null) {
-            eventHandler = context.getDebugSession().getEventHub().stepEvents().subscribe(debugEvent -> {
-                handleDebugEvent(debugEvent, context.getDebugSession(), context);
-            });
-        }
-
         switch (command) {
             case THREADS:
                 return this.threads((ThreadsArguments) arguments, response, context);
-            case STEPIN:
-                return this.stepIn((StepInArguments) arguments, response, context);
-            case STEPOUT:
-                return this.stepOut((StepOutArguments) arguments, response, context);
-            case NEXT:
-                return this.next((NextArguments) arguments, response, context);
             case PAUSE:
                 return this.pause((PauseArguments) arguments, response, context);
             case CONTINUE:
@@ -104,38 +76,6 @@ public class ThreadsRequestHandler implements IDebugRequestHandler {
             // when the thread is exiting.
         }
         response.body = new Responses.ThreadsResponseBody(threads);
-        return CompletableFuture.completedFuture(response);
-    }
-
-    private CompletableFuture<Response> stepIn(Requests.StepInArguments arguments, Response response, IDebugAdapterContext context) {
-        ThreadReference thread = DebugUtility.getThread(context.getDebugSession(), arguments.threadId);
-        if (thread != null) {
-            setPendingStepKind(thread, StepRequest.STEP_INTO);
-            setOriginalStackDepth(thread);
-            setOriginalStepLocation(thread);
-            DebugUtility.stepInto(thread, context.getDebugSession().getEventHub(), context.getDebugFilters().stepFilters);
-            checkThreadRunningAndRecycleIds(thread, context);
-        }
-        return CompletableFuture.completedFuture(response);
-    }
-
-    private CompletableFuture<Response> stepOut(Requests.StepOutArguments arguments, Response response, IDebugAdapterContext context) {
-        ThreadReference thread = DebugUtility.getThread(context.getDebugSession(), arguments.threadId);
-        if (thread != null) {
-            setPendingStepKind(thread, StepRequest.STEP_OUT);
-            DebugUtility.stepOut(thread, context.getDebugSession().getEventHub(), context.getDebugFilters().stepFilters);
-            checkThreadRunningAndRecycleIds(thread, context);
-        }
-        return CompletableFuture.completedFuture(response);
-    }
-
-    private CompletableFuture<Response> next(Requests.NextArguments arguments, Response response, IDebugAdapterContext context) {
-        ThreadReference thread = DebugUtility.getThread(context.getDebugSession(), arguments.threadId);
-        if (thread != null) {
-            setPendingStepKind(thread, StepRequest.STEP_OVER);
-            DebugUtility.stepOver(thread, context.getDebugSession().getEventHub(), context.getDebugFilters().stepFilters);
-            checkThreadRunningAndRecycleIds(thread, context);
-        }
         return CompletableFuture.completedFuture(response);
     }
 
@@ -175,7 +115,10 @@ public class ThreadsRequestHandler implements IDebugRequestHandler {
         return CompletableFuture.completedFuture(response);
     }
 
-    private void checkThreadRunningAndRecycleIds(ThreadReference thread, IDebugAdapterContext context) {
+    /**
+     * Recycle the related ids owned by the specified thread.
+     */
+    public static void checkThreadRunningAndRecycleIds(ThreadReference thread, IDebugAdapterContext context) {
         try {
             boolean allThreadsRunning = !DebugUtility.getAllThreadsSafely(context.getDebugSession()).stream()
                     .anyMatch(ThreadReference::isSuspended);
@@ -191,142 +134,5 @@ public class ThreadsRequestHandler implements IDebugRequestHandler {
             // isSuspended may throw ObjectCollectedException when the thread terminates
             context.getRecyclableIdPool().removeObjectsByOwner(thread.uniqueID());
         }
-    }
-
-    private void handleDebugEvent(DebugEvent debugEvent, IDebugSession debugSession, IDebugAdapterContext context) {
-        StepEvent event = (StepEvent) debugEvent.event;
-        ThreadReference thread = event.thread();
-        debugEvent.shouldResume = false;
-        if (context.getDebugFilters() != null) {
-            if (getPendingStepKind(thread) == StepRequest.STEP_INTO) {
-                // Check if the step into operation stepped through the filtered code and stopped at an un-filtered location.
-                if (getOriginalStackDepth(thread) + 1 < DebugUtility.getFrameCount(thread)) {
-                    // Create another stepOut request to return back where we started the step into.
-                    DebugUtility.stepOut(thread, debugSession.getEventHub(), context.getDebugFilters().stepFilters);
-                    return;
-                }
-                // If the ending step location is filtered, or same as the original location where the step into operation is originated,
-                // do another step of the same kind.
-                if (locationShouldBeFiltered(thread, context) || shouldDoExtraStepInto(thread)) {
-                    DebugUtility.stepInto(thread, debugSession.getEventHub(), context.getDebugFilters().stepFilters);
-                    return;
-                }
-            }
-        }
-        context.sendEvent(new Events.StoppedEvent("step", thread.uniqueID()));
-    }
-
-    /**
-     * Return true if the StepEvent's location is a Method that the user has indicated to filter.
-     */
-    private boolean locationShouldBeFiltered(ThreadReference thread, IDebugAdapterContext context) {
-        Location originalLocation = getOriginalStepLocation(thread);
-        Location currentLocation = null;
-        StackFrame topFrame = DebugUtility.getTopFrame(thread);
-        if (topFrame != null) {
-            currentLocation = topFrame.location();
-        }
-        if (originalLocation == null || currentLocation == null) {
-            return false;
-        }
-        return !methodShouldBeFiltered(originalLocation.method(), context) && methodShouldBeFiltered(currentLocation.method(), context);
-    }
-
-    private boolean methodShouldBeFiltered(Method method, IDebugAdapterContext context) {
-        if ((context.getDebugFilters().skipStaticInitializers && method.isStaticInitializer())
-                || (context.getDebugFilters().skipSynthetics && method.isSynthetic())
-                || (context.getDebugFilters().skipConstructors && method.isConstructor())) {
-            return true;
-        }
-        return false;
-    }
-
-    /**
-     * Check if the current top stack is same as the original top stack.
-     */
-    private boolean shouldDoExtraStepInto(ThreadReference thread) {
-        if (getOriginalStackDepth(thread) != DebugUtility.getFrameCount(thread)) {
-            return false;
-        }
-        Location originalLocation = getOriginalStepLocation(thread);
-        if (originalLocation == null) {
-            return false;
-        }
-        Location currentLocation = DebugUtility.getTopFrame(thread).location();
-        Method originalMethod = originalLocation.method();
-        Method currentMethod = currentLocation.method();
-        if (!originalMethod.equals(currentMethod)) {
-            return false;
-        }
-        if (originalLocation.lineNumber() != currentLocation.lineNumber()) {
-            return false;
-        }
-        return true;
-    }
-
-    private ThreadState getThreadState(ThreadReference thread) {
-        if (thread  == null) {
-            return null;
-        }
-        long threadId = thread.uniqueID();
-        if (!threadStates.containsKey(threadId)) {
-            threadStates.put(threadId, new ThreadState());
-        }
-        return threadStates.get(threadId);
-    }
-
-    private void setPendingStepKind(ThreadReference thread, int kind) {
-        ThreadState state = getThreadState(thread);
-        if (state != null) {
-            state.pendingStepKind = kind;
-        }
-    }
-
-    private int getPendingStepKind(ThreadReference thread) {
-        ThreadState state = getThreadState(thread);
-        if (state == null) {
-            return -1;
-        }
-        return state.pendingStepKind;
-    }
-
-    private void setOriginalStackDepth(ThreadReference thread) {
-        ThreadState state = getThreadState(thread);
-        if (state != null) {
-            state.stackDepth = DebugUtility.getFrameCount(thread);
-        }
-    }
-
-    private int getOriginalStackDepth(ThreadReference thread) {
-        ThreadState state = getThreadState(thread);
-        if (state == null) {
-            return -1;
-        }
-        return state.stackDepth;
-    }
-
-    private void setOriginalStepLocation(ThreadReference thread) {
-        ThreadState state = getThreadState(thread);
-        if (state != null) {
-            StackFrame topFrame = DebugUtility.getTopFrame(thread);
-            if (topFrame != null) {
-                state.stepLocation = topFrame.location();
-            }
-        }
-    }
-
-    private Location getOriginalStepLocation(ThreadReference thread) {
-        ThreadState state = getThreadState(thread);
-        if (state != null) {
-            return state.stepLocation;
-        }
-        return null;
-    }
-
-    class ThreadState {
-        int pendingStepKind = -1;
-        StepRequest pendingStepRequest = null;
-        int stackDepth = -1;
-        Location stepLocation = null;
     }
 }
